Updated dropdown and Primary Nav and Special logo so that they are all basically looking good.
